+++ src/handlers/hotelsHandlers.js
@@ -1,4 +1,4 @@
-const { getHotelById } = require('../controllers/hotelsControllers');
+const { getHotelById, postHotel } = require('../controllers/hotelsControllers.js');
 
 const getHotelIdHandler = async (req, res) => {
     const { id } = req.params;
@@ -11,6 +11,19 @@ const getHotelIdHandler = async (req, res) => {
     }
 }
 
+const postHotelHandler = async (req, res)=>{
+    //Debe recibir en el body todas las propiedades del hotel, junto con el id del pais
+    const hotelData = req.body
+
+    try{
+        const createdHotel = await postHotel(hotelData)
+        res.status(200).json(createdHotel)
+    } catch(error) {
+        res.status(500).json({error: error.message})
+    }
+}
+
 module.exports = {
-    getHotelIdHandler
+    getHotelIdHandler,
+    postHotelHandler
 }
