+++ src/routes/hotelsRouter.js
@@ -1,13 +1,10 @@
 const { Router } = require('express');
 const hotelsRouter = Router();
-const { getHotelIdHandler } = require('../handlers/hotelsHandlers');
-//Handlers para cada endpoint
-const { postHotelHandler } = require('../handlers/hotelHandlers')
+const { getHotelIdHandler, postHotelHandler } = require('../handlers/hotelsHandlers');
 
 //Endpoints
 //hotelsRouter.get("/", getHotels)
 hotelsRouter.post('/', postHotelHandler)
-
 
 hotelsRouter.get('/:id', getHotelIdHandler);
